rag-tools/requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
pytest>=8.2.0,<9.0.0
langchain==0.1.12
langchain-text-splitters==0.0.1 